Update StaffPicks.js, WhatsNewList.js and the More From list files
Added a line so that when you click on a a record from the home screen, that record is not included under the "More from..." lists in the details screen

The "More from..." lists now update if you clicik on a record - BUT only the specific "More from..." list from where you clicked on a record. i.e. if you click on a record under "More from this artist", only that list updates, not the "more from this label" list